web/requirements.txt
elasticsearch==7.9.1
dash
plotly
pandas
requests
kaleido
cookiecutter
dash-bootstrap-components
Pillow
dash-mantine-components
python-dotenv